Add ui/src/[images,components,css] and move files into them
