+++ ui/src/components/Interapp.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-import './App.css';
-import './main.css';
+import '../css/App.css';
+import '../css/main.css';
 
 // This class defines teh apps loaded in each section of the landing page
 // To be populated by storybook, and then later by the APIs
